==> 0_preprocess_data.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split

input_data = pd.read_csv('data/credit_score_data.csv')

x, y = input_data.iloc[:, :-1], input_data.iloc[:, -1]
x_train, x_test, y_train, y_test = train_test_split(x, y, test_size=0.2, random_state=42)

train_full = pd.concat([x_train, y_train], axis=1)
test_full = pd.concat([x_test, y_test], axis=1)

train_full.to_csv("data/train.csv", index=False)
train_full

test_full.to_csv("data/test.csv", index=False)
test_full
