==> 1_notebook (description statistics).ipynb <==
import pandas as pd
import seaborn as sns
from matplotlib import pyplot as plt

def describe_data(data):
    print("Data description:")
    for _, column_name in enumerate(data):
        print(data[column_name].describe())
        print(f"Number of NaN values: {data[column_name].isna().sum()}")
        print("=====================================")

def plot_distributions(input_data):
    for column in input_data.columns:
        plt.figure(figsize=(15, 10))

        if input_data[column].dtype == "int64" or input_data[column].dtype == "float64":
            ax = sns.histplot(input_data[column], kde=True)
            ax.set(xlabel=column, ylabel='Count')
            plt.title(f"Distribution of {column}")
            plt.show()
        elif len(input_data[column].unique()) < 50:
            ax = sns.barplot(x = input_data[column].value_counts().index, y = input_data[column].value_counts())
            ax.set(xlabel=column, ylabel='Count')
            ax.set_xticklabels(ax.get_xticklabels(), rotation=55)
            plt.title(f"Distribution of {column}")
            plt.show()
        else:
            ax = sns.barplot(x = input_data[column].value_counts().head(50).index, y = input_data[column].value_counts().head(50))
            ax.set(xlabel=column, ylabel='Count')
            ax.set_xticklabels(ax.get_xticklabels(), rotation=55)
            plt.title(f"Distribution of top 50 {column} values")
            plt.show()

input_data = pd.read_csv("data/train.csv", encoding="UTF-8")
describe_data(input_data)

plot_distributions(input_data)
