==> 3_notebook_features_extraction.ipynb <==
import pandas as pd

train_data = pd.read_csv("data/train_data_preprocessed.csv")
test_data = pd.read_csv("data/test_data_preprocessed.csv")

train_y = train_data["Credit_Score"]
train_x = train_data.drop(["Credit_Score"], axis=1)
train_x

from sklearn.model_selection import cross_val_score
from sklearn.linear_model import LogisticRegression

model = LogisticRegression(max_iter=10000, random_state=42)
print(cross_val_score(model, train_x, train_y, scoring="f1_macro", n_jobs=-1).mean())

from sklearn.feature_selection import f_classif

train_features = train_x.keys()
f_statistic, p_values = f_classif(train_x, train_y)
f_statistic

from pandas import Index

corr_bad_columns = list(Index(['Credit_Utilization_Ratio',
                          'Occupation_Accountant', 'Occupation_Architect', 'Occupation_Developer',
                          'Occupation_Doctor', 'Occupation_Engineer', 'Occupation_Entrepreneur',
                          'Occupation_Journalist', 'Occupation_Lawyer', 'Occupation_Manager',
                          'Occupation_Mechanic', 'Occupation_Media_Manager',
                          'Occupation_Musician', 'Occupation_Scientist', 'Occupation_Teacher',
                          'Occupation_Writer', 'Occupation________', 'Age_categories_Youth',
                          'Age_categories_Young_Adult', 'Age_categories_Adult',
                          'Credit_Mix_Standard', 'Credit_Mix__', 'Payment_of_Min_Amount_NM',
                          'Payment_Behaviour_!@9#%8',
                          'Payment_Behaviour_High_spent_Large_value_payments',
                          'Payment_Behaviour_High_spent_Medium_value_payments',
                          'Payment_Behaviour_High_spent_Small_value_payments',
                          'Payment_Behaviour_Low_spent_Large_value_payments',
                          'Payment_Behaviour_Low_spent_Medium_value_payments',
                          'Payment_Behaviour_Low_spent_Small_value_payments'],
                         dtype='object'))

columns_to_delete = pd.Series(f_statistic).where(f_statistic < 300).dropna().index
f_score_bad_columns = list(train_features[columns_to_delete])
f_score_bad_columns

bad_columns = list(set(f_score_bad_columns + corr_bad_columns))
print(f"Features number before deletion: {train_data.shape[1]}")
print(f"Features number to delete: {len(bad_columns)}")
bad_columns

train_data = train_data.drop(bad_columns, axis=1)
test_data = test_data.drop(bad_columns, axis=1)

train_data

test_data

y = train_data["Credit_Score"]
x = train_data.drop(["Credit_Score"], axis=1)

print(cross_val_score(model, x, y, scoring="f1_macro", n_jobs=-1).mean())

train_data.to_csv("data/train_data_features_selected.csv", index=False)
test_data.to_csv("data/test_data_features_selected.csv", index=False)

from sklearn.feature_selection import RFECV

rfecv = RFECV(
    estimator=model,
    scoring="f1_macro",
    n_jobs=-1,
    min_features_to_select=10
)
rfecv.fit(train_x, train_y)
print(f"Optimal number of features: {rfecv.n_features_}")
rfecv.support_

rfecv.support_

rfecv_good_columns = list(train_x.keys()[rfecv.support_])
print(cross_val_score(model, train_x[rfecv_good_columns], train_y, scoring="f1_macro", n_jobs=-1).mean())

train_data = pd.read_csv("data/train_data_preprocessed.csv")
test_data = pd.read_csv("data/test_data_preprocessed.csv")

train_data = train_data[rfecv_good_columns + ["Credit_Score"]]
test_data = test_data[rfecv_good_columns + ["Credit_Score"]]

train_data

test_data

train_data.to_csv("data/train_data_features_selected.csv", index=False)
test_data.to_csv("data/test_data_features_selected.csv", index=False)

train_data.drop(["Credit_Score"], axis=1).keys()
