==> 4_notebook_balancing.ipynb <==
import pandas as pd

train = pd.read_csv("data/train_data_features_selected.csv")
test = pd.read_csv("data/test_data_features_selected.csv")
train

test

print(train["Credit_Score"].value_counts())

poor_num_to_add = train["Credit_Score"].value_counts()["Standard"] - train["Credit_Score"].value_counts()["Poor"]
print(poor_num_to_add)
good_num_to_add = train["Credit_Score"].value_counts()["Standard"] - train["Credit_Score"].value_counts()["Good"]
print(good_num_to_add)

from sklearn.utils import resample

train_standard = train[train["Credit_Score"] == "Standard"]
print(train_standard.shape[0])

train_poor = train[train["Credit_Score"] == "Poor"]
print(train_poor.shape[0])
train_good = train[train["Credit_Score"] == "Good"]
print(train_good.shape[0])


train_poor_upsampled = resample(train_poor,random_state=42, n_samples=train_poor.shape[0]+poor_num_to_add, replace=True)
print(train_poor_upsampled.shape[0])
train_good_upsampled = resample(train_good,random_state=42, n_samples=train_good.shape[0]+good_num_to_add, replace=True)
print(train_good_upsampled.shape[0])

train_upsampled = pd.concat([train_standard, train_poor_upsampled, train_good_upsampled])
train_upsampled

from sklearn.model_selection import cross_val_score
from sklearn.linear_model import LogisticRegression

train_y = train_upsampled["Credit_Score"]
train_x = train_upsampled.drop("Credit_Score", axis=1)

model = LogisticRegression(max_iter=10000, random_state=42)
print(cross_val_score(model, train_x, train_y, scoring="f1_macro", n_jobs=-1).mean())

print("Old f1 score (before upsampling):")
print(cross_val_score(model, train.drop("Credit_Score", axis=1), train["Credit_Score"], scoring="f1_macro", n_jobs=-1).mean())

train_upsampled["Credit_Score"].value_counts(normalize=True)

train_upsampled = train_upsampled.reset_index(drop=True)
train_upsampled

train_upsampled.to_csv("data/train_data_features_selected_upsampled.csv", index=False)
