==> 5_notebook_cv.ipynb <==
import pandas as pd

train = pd.read_csv('data/train_data_features_selected_upsampled.csv')
test = pd.read_csv('data/test_data_features_selected.csv')

train["Credit_Score"].value_counts(normalize=True)

test["Credit_Score"].value_counts(normalize=True)

train_y = train["Credit_Score"]
train_x = train.drop(["Credit_Score"], axis=1)

test_y = test["Credit_Score"]
test_x = test.drop(["Credit_Score"], axis=1)

train_x

train_y

from sklearn.linear_model import LogisticRegression
from sklearn.preprocessing import StandardScaler, MinMaxScaler
from sklearn.pipeline import Pipeline

cv_number = 3
penalties = ['l1', 'l2', 'none']
c_values = [1, 10, 50]
random_state = [42]
max_features = ['sqrt', 'log2']
n_estimators = [10, 50, 100]

methods_stats_df = pd.DataFrame(data=None, columns=['method', 'f1_macro', 'scaler', 'params'])
methods_stats_df_index = 0

def append_method_stats(method, accuracy, scaler, params):
    global methods_stats_df_index
    methods_stats_df.loc[methods_stats_df_index] = [method, accuracy, scaler, params]
    methods_stats_df_index += 1

standard_scaler_pipeline = Pipeline([('scaler', StandardScaler()),
                                     ('logreg', LogisticRegression())])

min_max_scaler_pipeline = Pipeline([('scaler', MinMaxScaler()),
                                    ('logreg', LogisticRegression())])

param_grid = {
    'logreg__penalty': penalties,
    'logreg__C': c_values,
    'logreg__random_state': random_state
}

from sklearn.model_selection import GridSearchCV

gridsearch = GridSearchCV(standard_scaler_pipeline,
                          param_grid=param_grid,
                          scoring='f1_macro',
                          cv=cv_number,
                          n_jobs=-1)
gridsearch.fit(train_x, train_y)

append_method_stats('LogisticRegression', gridsearch.best_score_, 'StandardScaler', gridsearch.best_params_)

gridsearch = GridSearchCV(min_max_scaler_pipeline,
                          param_grid=param_grid,
                          scoring='f1_macro',
                          cv=cv_number,
                          n_jobs=-1)
gridsearch.fit(train_x, train_y)

append_method_stats('LogisticRegression', gridsearch.best_score_, 'MinMaxScaler', gridsearch.best_params_)

from sklearn.svm import SVC

standard_scaler_pipeline = Pipeline([('scaler', StandardScaler()),
                                     ('svc', SVC())])

min_max_scaler_pipeline = Pipeline([('scaler', MinMaxScaler()),
                                    ('svc', SVC())])

param_grid = {
    'svc__C': c_values,
    'svc__random_state': random_state
}

gridsearch = GridSearchCV(standard_scaler_pipeline,
                          param_grid=param_grid,
                          scoring='f1_macro',
                          cv=cv_number,
                          n_jobs=-1)
gridsearch.fit(train_x, train_y)

append_method_stats('SVC', gridsearch.best_score_, 'StandardScaler', gridsearch.best_params_)

gridsearch = GridSearchCV(min_max_scaler_pipeline,
                          param_grid=param_grid,
                          scoring='f1_macro',
                          cv=cv_number,
                          n_jobs=-1)
gridsearch.fit(train_x, train_y)

append_method_stats('SVC', gridsearch.best_score_, 'MinMaxScaler', gridsearch.best_params_)

from sklearn.neighbors import KNeighborsClassifier

standard_scaler_pipeline = Pipeline([('scaler', StandardScaler()),
                                     ('knn', KNeighborsClassifier())])

min_max_scaler_pipeline = Pipeline([('scaler', MinMaxScaler()),
                                    ('knn', KNeighborsClassifier())])

param_grid = {
    'knn__n_neighbors': [2, 3, 5, 7, 10],
    'knn__weights': ['uniform', 'distance'],
    'knn__algorithm': ['auto', 'ball_tree', 'kd_tree', 'brute'],
    'knn__p': [1, 2]
}

gridsearch = GridSearchCV(standard_scaler_pipeline,
                          param_grid=param_grid,
                          scoring='f1_macro',
                          cv=cv_number,
                          n_jobs=-1)
gridsearch.fit(train_x, train_y)

append_method_stats('KNeighborsClassifier', gridsearch.best_score_, 'StandardScaler', gridsearch.best_params_)

gridsearch = GridSearchCV(min_max_scaler_pipeline,
                          param_grid=param_grid,
                          scoring='f1_macro',
                          cv=cv_number,
                          n_jobs=-1)
gridsearch.fit(train_x, train_y)

append_method_stats('KNeighborsClassifier', gridsearch.best_score_, 'MinMaxScaler', gridsearch.best_params_)

from sklearn.naive_bayes import MultinomialNB

min_max_scaler_pipeline = Pipeline([('scaler', MinMaxScaler()),
                                    ('nb', MultinomialNB())])

param_grid = {
    'nb__alpha': [0, 0.5, 1.0, 1.5, 2.0],
    'nb__fit_prior': [True, False]
}

gridsearch = GridSearchCV(min_max_scaler_pipeline,
                          param_grid=param_grid,
                          scoring='f1_macro',
                          cv=cv_number,
                          n_jobs=-1)
gridsearch.fit(train_x, train_y)

append_method_stats('MultinomialNB', gridsearch.best_score_, 'MinMaxScaler', gridsearch.best_params_)

from sklearn.tree import DecisionTreeClassifier

standard_scaler_pipeline = Pipeline([('scaler', StandardScaler()),
                                     ('dt', DecisionTreeClassifier())])

min_max_scaler_pipeline = Pipeline([('scaler', MinMaxScaler()),
                                    ('dt', DecisionTreeClassifier())])

param_grid = {
    'dt__criterion': ['gini', 'entropy', 'log_loss'],
    'dt__splitter': ['best', 'random'],
    'dt__max_depth': [None, 10, 20],
    'dt__min_samples_split': [2, 4, 6],
    'dt__min_samples_leaf': [1, 2, 3],
    'dt__max_features': max_features,
    'dt__random_state': random_state
}

gridsearch = GridSearchCV(standard_scaler_pipeline,
                          param_grid=param_grid,
                          scoring='f1_macro',
                          cv=cv_number,
                          n_jobs=-1)
gridsearch.fit(train_x, train_y)

append_method_stats('DecisionTreeClassifier', gridsearch.best_score_, 'StandardScaler', gridsearch.best_params_)

gridsearch = GridSearchCV(min_max_scaler_pipeline,
                          param_grid=param_grid,
                          scoring='f1_macro',
                          cv=cv_number,
                          n_jobs=-1)
gridsearch.fit(train_x, train_y)

append_method_stats('DecisionTreeClassifier', gridsearch.best_score_, 'MinMaxScaler', gridsearch.best_params_)

from sklearn.ensemble import RandomForestClassifier

standard_scaler_pipeline = Pipeline([('scaler', StandardScaler()),
                                     ('rf', RandomForestClassifier())])

min_max_scaler_pipeline = Pipeline([('scaler', MinMaxScaler()),
                                    ('rf', RandomForestClassifier())])

param_grid = {
    'rf__criterion': ['gini', 'entropy'],
    'rf__n_estimators': n_estimators,
    'rf__min_samples_split': [2, 4, 6],
    'rf__min_samples_leaf': [1, 2, 3],
    'rf__max_features': max_features,
    'rf__random_state': random_state
}

gridsearch = GridSearchCV(standard_scaler_pipeline,
                          param_grid=param_grid,
                          scoring='f1_macro',
                          cv=cv_number,
                          n_jobs=-1)
gridsearch.fit(train_x, train_y)

append_method_stats('RandomForestClassifier', gridsearch.best_score_, 'StandardScaler', gridsearch.best_params_)

gridsearch = GridSearchCV(min_max_scaler_pipeline,
                          param_grid=param_grid,
                          scoring='f1_macro',
                          cv=cv_number,
                          n_jobs=-1)
gridsearch.fit(train_x, train_y)

append_method_stats('RandomForestClassifier', gridsearch.best_score_, 'MinMaxScaler', gridsearch.best_params_)

from sklearn.ensemble import GradientBoostingClassifier

standard_scaler_pipeline = Pipeline([('scaler', StandardScaler()),
                                     ('gb', GradientBoostingClassifier())])

min_max_scaler_pipeline = Pipeline([('scaler', MinMaxScaler()),
                                    ('gb', GradientBoostingClassifier())])

param_grid = {
    'gb__learning_rate': [0.1, 0.5, 1.0, 2.0, 3.0],
    'gb__max_features': max_features,
    'gb__random_state': random_state
}

gridsearch = GridSearchCV(standard_scaler_pipeline,
                          param_grid=param_grid,
                          scoring='f1_macro',
                          cv=cv_number,
                          n_jobs=-1)
gridsearch.fit(train_x, train_y)

append_method_stats('GradientBoostingClassifier', gridsearch.best_score_, 'StandardScaler', gridsearch.best_params_)

gridsearch = GridSearchCV(min_max_scaler_pipeline,
                          param_grid=param_grid,
                          scoring='f1_macro',
                          cv=cv_number,
                          n_jobs=-1)
gridsearch.fit(train_x, train_y)

append_method_stats('GradientBoostingClassifier', gridsearch.best_score_, 'MinMaxScaler', gridsearch.best_params_)

from sklearn.ensemble import BaggingClassifier

standard_scaler_pipeline = Pipeline([('scaler', StandardScaler()),
                                     ('bag', BaggingClassifier())])

min_max_scaler_pipeline = Pipeline([('scaler', MinMaxScaler()),
                                    ('bag', BaggingClassifier())])

param_grid = {
    'bag__n_estimators': n_estimators,
    'bag__random_state': random_state
}

gridsearch = GridSearchCV(standard_scaler_pipeline,
                          param_grid=param_grid,
                          scoring='f1_macro',
                          cv=cv_number,
                          n_jobs=-1)
gridsearch.fit(train_x, train_y)

append_method_stats('BaggingClassifier', gridsearch.best_score_, 'StandardScaler', gridsearch.best_params_)

gridsearch = GridSearchCV(min_max_scaler_pipeline,
                          param_grid=param_grid,
                          scoring='f1_macro',
                          cv=cv_number,
                          n_jobs=-1)
gridsearch.fit(train_x, train_y)

append_method_stats('BaggingClassifier', gridsearch.best_score_, 'MinMaxScaler', gridsearch.best_params_)

methods_stats_df.to_csv('stats/methods_stats.csv', index=False)
