==> 6_notebook_tests.ipynb <==
import pandas as pd
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn.preprocessing import MinMaxScaler, StandardScaler
from sklearn.tree import DecisionTreeClassifier

random_state = 42
cv_number = 3

train = pd.read_csv('data/train_data_features_selected_upsampled.csv')
test = pd.read_csv('data/test_data_features_selected.csv')

train_y = train["Credit_Score"]
train_x = train.drop(["Credit_Score"], axis=1)

test_y = test["Credit_Score"]
test_x = test.drop(["Credit_Score"], axis=1)

test_x_without_nan = test_x.fillna(0)

standard_scaled_train_x = StandardScaler().fit_transform(train_x)
standard_scaled_test_x = StandardScaler().fit_transform(test_x_without_nan)

minmax_scaled_train_x = MinMaxScaler().fit_transform(train_x)
minmax_scaled_test_x = MinMaxScaler().fit_transform(test_x_without_nan)

from sklearn.ensemble import BaggingClassifier
from sklearn.metrics import classification_report

bagging = BaggingClassifier(n_estimators=100, random_state=random_state)

bagging.fit(standard_scaled_train_x, train_y)
predicted_y = bagging.predict(standard_scaled_test_x)

print(classification_report(test_y, predicted_y))

bagging.fit(minmax_scaled_train_x, train_y)
predicted_y = bagging.predict(minmax_scaled_test_x)

print(classification_report(test_y, predicted_y))

random_forest = RandomForestClassifier(criterion='gini', max_features='sqrt', min_samples_leaf=1, min_samples_split=2, n_estimators=100, random_state=random_state)

random_forest.fit(standard_scaled_train_x, train_y)
predicted_y = random_forest.predict(standard_scaled_test_x)

print(classification_report(test_y, predicted_y))

random_forest.fit(minmax_scaled_train_x, train_y)
predicted_y = random_forest.predict(minmax_scaled_test_x)

print(classification_report(test_y, predicted_y))

knn_model = KNeighborsClassifier(algorithm='auto', n_neighbors=2, p=1, weights='distance')

knn_model.fit(standard_scaled_train_x, train_y)
predicted_y = knn_model.predict(standard_scaled_test_x)

print(classification_report(test_y, predicted_y))

knn_model.fit(minmax_scaled_train_x, train_y)
predicted_y = knn_model.predict(minmax_scaled_test_x)

print(classification_report(test_y, predicted_y))

d_tree_model = DecisionTreeClassifier(criterion='entropy', max_depth=None, max_features='sqrt', min_samples_leaf=1, min_samples_split=2, random_state=random_state, splitter='best')

d_tree_model.fit(minmax_scaled_train_x, train_y)
predicted_y = d_tree_model.predict(minmax_scaled_test_x)

print(classification_report(test_y, predicted_y))

d_tree_model.fit(standard_scaled_train_x, train_y)
predicted_y = d_tree_model.predict(standard_scaled_test_x)

print(classification_report(test_y, predicted_y))

from sklearn.svm import SVC

svc = SVC(C=50, random_state=random_state)

svc.fit(standard_scaled_train_x, train_y)
predicted_y = svc.predict(standard_scaled_test_x)

print(classification_report(test_y, predicted_y))

svc.fit(minmax_scaled_train_x, train_y)
predicted_y = svc.predict(minmax_scaled_test_x)

print(classification_report(test_y, predicted_y))

g_boosting = GradientBoostingClassifier(learning_rate=1.0, max_features='sqrt', random_state=random_state)

g_boosting.fit(standard_scaled_train_x, train_y)
predicted_y = g_boosting.predict(standard_scaled_test_x)

print(classification_report(test_y, predicted_y))

g_boosting.fit(minmax_scaled_train_x, train_y)
predicted_y = g_boosting.predict(minmax_scaled_test_x)

print(classification_report(test_y, predicted_y))

from sklearn.linear_model import LogisticRegression

logistic_regression = LogisticRegression(C=1, penalty='none', random_state=random_state)

logistic_regression.fit(minmax_scaled_train_x, train_y)
predicted_y = logistic_regression.predict(minmax_scaled_test_x)

print(classification_report(test_y, predicted_y))

logistic_regression = LogisticRegression(C=1, penalty='l2', random_state=random_state)

logistic_regression.fit(standard_scaled_train_x, train_y)
predicted_y = logistic_regression.predict(standard_scaled_test_x)

print(classification_report(test_y, predicted_y))

from sklearn.naive_bayes import MultinomialNB

nb_clf = MultinomialNB(alpha=2.0, fit_prior=True)

nb_clf.fit(minmax_scaled_train_x, train_y)
predicted_y = nb_clf.predict(minmax_scaled_test_x)

print(classification_report(test_y, predicted_y))
